move ConditionWriter out of Utils
