+++ inject-generator/src/main/java/io/avaje/inject/generator/MethodReader.java
@@ -276,7 +276,7 @@ final class MethodReader {
   }
 
   void buildConditional(Append writer) {
-    Util.buildBeanConditional(writer, conditions);
+    new ConditionalWriter(writer, conditions).write();
   }
 
   void buildAddFor(Append writer) {
