+++ inject-generator/src/main/java/io/avaje/inject/generator/Util.java
@@ -248,64 +248,4 @@ final class Util {
       .replace(Constants.DI, "");
   }
 
-  public static void buildBeanConditional(Append writer, BeanConditions conditions) {
-    if (conditions.isEmpty()) {
-      return;
-    }
-    new ConditionalWriter(writer, conditions).write();
-  }
-
-  private static final class ConditionalWriter {
-    private final Append writer;
-    private final BeanConditions conditions;
-
-    private boolean first = true;
-
-    private ConditionalWriter(Append writer, BeanConditions conditions) {
-      this.writer = writer;
-      this.conditions = conditions;
-    }
-
-    public void write() {
-      writer.append("    if (");
-
-      for (final var requireType : conditions.requireTypes) {
-        prefix();
-        writer.append("!builder.contains(%s.class)", Util.shortName(requireType));
-      }
-      for (final var missing : conditions.missingTypes) {
-        prefix();
-        writer.append("builder.contains(%s.class)", Util.shortName(missing));
-      }
-      for (final var qualifier : conditions.qualifierNames) {
-        prefix();
-        writer.append("!builder.containsQualifier(\"%s\")", qualifier);
-      }
-      for (final var props : conditions.containsProps) {
-        prefix();
-        writer.append("builder.property().missing(\"%s\")", props);
-      }
-      for (final var props : conditions.missingProps) {
-        prefix();
-        writer.append("builder.property().contains(\"%s\")", props);
-      }
-      for (final var props : conditions.propertyEquals.entrySet()) {
-        prefix();
-        writer.append("builder.property().notEqualTo(\"%s\", \"%s\")", props.getKey(), props.getValue());
-      }
-      for (final var props : conditions.propertyNotEquals.entrySet()) {
-        prefix();
-        writer.append("builder.property().equalTo(\"%s\", \"%s\")", props.getKey(), props.getValue());
-      }
-      writer.append(") {").eol().append("      return;").eol().append("    }").eol().eol();
-    }
-
-    private void prefix() {
-      if (first) {
-        first = false;
-      } else {
-        writer.eol().append("      || ");
-      }
-    }
-  }
 }
